Send IdentifyMessage to client to set Points&Prio.
Immediately after connecting, the server will now send
a IdentifyMessage to the client. This message contains
nothing more than the snake's id. This id is then set
at the client's game. By this the client knows which
snake belongs to him/her and so the current priority
and points can be set accordingly.